// src/webview/uxp-api/modules/photoshop/types.ts
export interface ImagingBounds {
  readonly left: number;
  readonly right: number;
  readonly top: number;
  readonly bottom: number;
  readonly width: number;
  readonly height: number;
}